Reorganise types and modules.
