+++ src/daemon/mod.rs
@@ -5,3 +5,4 @@ pub mod http;
 pub mod krillserver;
 pub mod mq;
 pub mod scheduler;
+pub mod ta;
